+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 -----
 
 * :class:`.AgateTestCase` is available as ``agate.testcase.AgateTestCase`` for extensions to use. (#384)
-* :meth:`.Table.reject` implemented. Opposite of :meth:`.Table.select`. (#388)
+* :meth:`.Table.exclude` implemented. Opposite of :meth:`.Table.select`. (#388)
 * :meth:`.Table.merge` now accepts a ``row_names`` argument. (#403)
 * :class:`.Formula` now automatically casts computed values to specified data type unless ``invalidate`` is set to ``False``. (#398)
 * Added Neil Bedi to AUTHORS.